Create defaultAssets.tsx
Update constants.tsx

Update CurrencyList.tsx
