+++ packages/extension-polkagate/src/util/constants.tsx
@@ -211,281 +211,3 @@ export const REGISTRARS_LIST: { addresses: string[]; index: number; name: string
 /** Login Password constants */
 export const NO_PASS_PERIOD = 30 * 60 * 1000; // in ms, the duration of time we do not ask user for password after a successful login
 export const MAYBE_LATER_PERIOD = 5 * 60 * 1000; // ms
-export const currencyList = [
-  {
-    code: 'USD',
-    country: 'United States',
-    currency: 'Dollar',
-    sign: '$'
-  },
-  {
-    code: 'AED',
-    country: 'Emirates',
-    currency: 'Dirham',
-    sign: 'DH'
-  },
-  {
-    code: 'ARS',
-    country: 'Argentina',
-    currency: 'Peso',
-    sign: '$'
-  },
-  {
-    code: 'AUD',
-    country: 'Australia',
-    currency: 'Australian dollar',
-    sign: 'A$'
-  },
-  {
-    code: 'BDT',
-    country: 'Bangladesh',
-    currency: 'Taka',
-    sign: '৳'
-  },
-  {
-    code: 'BHD',
-    country: 'Bahrein',
-    currency: 'Dinar',
-    sign: 'BD'
-  },
-  {
-    code: 'BMD',
-    country: 'Bermuda',
-    currency: 'Bermuda Dollar',
-    sign: '$'
-  },
-  {
-    code: 'BRL',
-    country: 'Brazil',
-    currency: 'Real',
-    sign: 'R$'
-  },
-  {
-    code: 'CAD',
-    country: 'Canada',
-    currency: 'Dollar',
-    sign: 'CA$'
-  },
-  {
-    code: 'CHF',
-    country: 'Swiss',
-    currency: 'Franc',
-    sign: 'Fr.'
-  },
-  {
-    code: 'CLP',
-    country: 'Chile',
-    currency: 'Peso',
-    sign: 'CLP$'
-  },
-  {
-    code: 'CNY',
-    country: 'China',
-    currency: 'Yuan',
-    sign: '¥'
-  },
-  {
-    code: 'CNY',
-    country: 'Czech',
-    currency: 'Koruna',
-    sign: 'CZK'
-  },
-  {
-    code: 'DKK',
-    country: 'Denmark',
-    currency: 'Krone',
-    sign: 'kr.'
-  },
-  {
-    code: 'EUR',
-    country: 'European Union',
-    currency: 'Euro',
-    sign: '€'
-  },
-  {
-    code: 'GBP',
-    country: 'Great British',
-    currency: 'Pound',
-    sign: '£'
-  },
-  {
-    code: 'GEL',
-    country: 'Georgia',
-    currency: 'Lari',
-    sign: '₾'
-  },
-  {
-    code: 'HKD',
-    country: 'Hong Kong',
-    currency: 'Dollar',
-    sign: '$'
-  },
-  {
-    code: 'HUF',
-    country: 'Hungary',
-    currency: 'Forint',
-    sign: 'Ft'
-  },
-  {
-    code: 'IDR',
-    country: 'Indonesia',
-    currency: 'Rupiah',
-    sign: 'Rp'
-  },
-  {
-    code: 'ILS',
-    country: 'Israeli',
-    currency: 'shekel',
-    sign: '₪'
-  },
-  {
-    code: 'INR',
-    country: 'India',
-    currency: 'Rupee',
-    sign: '₹'
-  },
-  {
-    code: 'JPY',
-    country: 'Japan',
-    currency: 'Yen',
-    sign: '¥'
-  },
-  {
-    code: 'KRW',
-    country: 'South Korea',
-    currency: 'Won',
-    sign: '₩'
-  },
-  {
-    code: 'KWD',
-    country: 'Kuwait',
-    currency: 'Dinar',
-    sign: 'KD'
-  },
-  {
-    code: 'LKR',
-    country: 'Sri Lanka',
-    currency: 'Rupee',
-    sign: 'Rs'
-  },
-  {
-    code: 'MMK',
-    country: 'Myanmar',
-    currency: 'Kyat',
-    sign: 'K'
-  },
-  {
-    code: 'MXN',
-    country: 'Mexico',
-    currency: 'Peso',
-    sign: 'MX$'
-  },
-  {
-    code: 'MYR',
-    country: 'Malaysia',
-    currency: 'Ringgit',
-    sign: 'RM'
-  },
-  {
-    code: 'NGN',
-    country: 'Nigeria',
-    currency: 'Naira',
-    sign: '₦'
-  },
-  {
-    code: 'NOK',
-    country: 'Norway',
-    currency: 'Krone',
-    sign: 'kr'
-  },
-  {
-    code: 'NZD',
-    country: 'New Zealand',
-    currency: 'Dollar',
-    sign: 'NZ$'
-  },
-  {
-    code: 'PHP',
-    country: 'Philippines',
-    currency: 'peso',
-    sign: '₱'
-  },
-  {
-    code: 'PKR',
-    country: 'Pakistan',
-    currency: 'Rupee',
-    sign: 'Rs'
-  },
-  {
-    code: 'PLN',
-    country: 'Poland',
-    currency: 'złoty',
-    sign: 'zł'
-  },
-  {
-    code: 'RUB',
-    country: 'Russia',
-    currency: 'Ruble',
-    sign: '₽'
-  },
-  {
-    code: 'SAR',
-    country: 'Saudi Arabia',
-    currency: 'Riyal',
-    sign: 'SR'
-  },
-  {
-    code: 'SEK',
-    country: 'Sweden',
-    currency: 'Krona',
-    sign: 'kr'
-  },
-  {
-    code: 'SGD',
-    country: 'Singapore',
-    currency: 'Dollar',
-    sign: 'S$'
-  },
-  {
-    code: 'THB',
-    country: 'Thailand',
-    currency: 'Baht',
-    sign: '฿'
-  },
-  {
-    code: 'TRY',
-    country: 'Turkey',
-    currency: 'lira',
-    sign: '₺'
-  },
-  {
-    code: 'TWD',
-    country: 'Taiwan',
-    currency: 'Dollar',
-    sign: 'NT$'
-  },
-  {
-    code: 'UAH',
-    country: 'Ukraine',
-    currency: 'hryvnia',
-    sign: '₴'
-  },
-  {
-    code: 'VEF',
-    country: 'Venezuela',
-    currency: 'Bolívar',
-    sign: 'Bs.F'
-  },
-  {
-    code: 'VND',
-    country: 'Vietnam',
-    currency: 'Dong',
-    sign: '₫'
-  },
-  {
-    code: 'ZAR',
-    country: 'South Africa',
-    currency: 'Rand',
-    sign: 'R'
-  }
-];
